requirements.txt
flask==2.3.3
pyserial==3.5
Werkzeug==2.3.7
Jinja2==3.1.2